github.com/FIS/terraform/functions/fis-telemetry-adx/requirements.txt
azure-core==1.24.1
azure-digitaltwins-core==1.1.0
azure-eventhub==5.9.0
azure-eventhub-checkpointstoreblob==1.1.4
azure-functions==1.11.2
azure-identity==1.10.0
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.2
idna==3.3
isodate==0.6.1
msal==1.18.0
msal-extensions==1.0.0
msrest==0.6.21
oauthlib==3.2.0
portalocker==2.4.0
pycparser==2.21
PyJWT==2.4.0
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
typing-extensions==4.2.0
uamqp==1.5.3
urllib3==1.26.9
